+++ example/integration_test/platform_test.dart
@@ -8,14 +8,31 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
 import 'lib/lib_source_test_data.dart';
-import 'lib/lib_test_utils.dart';
 import 'platform_features.dart';
+import 'source_test_data.dart';
 import 'test_utils.dart';
+
+const _defaultTimeout = Duration(seconds: 30);
+
+final isLinux = !kIsWeb && defaultTargetPlatform == TargetPlatform.linux;
+final isAndroid = !kIsWeb && defaultTargetPlatform == TargetPlatform.android;
+
+bool canDetermineDuration(SourceTestData td) {
+  // TODO(gustl22): cannot determine duration for VBR on Linux
+  // FIXME(gustl22): duration event is not emitted for short duration
+  // WAV on Linux (only platform tests, may be a race condition).
+  if (td.duration == null) {
+    return true;
+  }
+  if (isLinux) {
+    return !(td.isVBR || td.duration! < const Duration(seconds: 5));
+  }
+  return true;
+}
 
 void main() async {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
   final features = PlatformFeatures.instance();
-  final isLinux = !kIsWeb && defaultTargetPlatform == TargetPlatform.linux;
   final audioTestDataList = await getAudioTestDataList();
 
   group('Platform method channel', () {
@@ -35,42 +52,45 @@ void main() async {
     testWidgets(
       'Throw PlatformException, when loading invalid file',
       (tester) async {
-        try {
-          // Throws PlatformException via MethodChannel:
-          await tester.prepareSource(
+        // Throws PlatformException instead of returning prepared event.
+        await tester.expectSettingSourceFailure(
+          future: tester.prepareSource(
             playerId: playerId,
             platform: platform,
             testData: invalidAssetTestData,
-          );
-          fail('PlatformException not thrown');
-        } on PlatformException catch (e) {
-          expect(e.message, startsWith('Failed to set source.'));
+          ),
+        );
+
+        if (isLinux) {
+          // Linux throws a second failure event for invalid files.
+          // If not caught, it would be randomly thrown in the following tests.
+          final nextEvent = platform.getEventStream(playerId).first;
+          await tester.expectSettingSourceFailure(future: nextEvent);
         }
-        await tester.pumpLinux();
       },
     );
 
     testWidgets(
       'Throw PlatformException, when loading non existent file',
       (tester) async {
-        try {
-          // Throws PlatformException via MethodChannel:
-          await tester.prepareSource(
+        // Throws PlatformException instead of returning prepared event.
+        await tester.expectSettingSourceFailure(
+          future: tester.prepareSource(
             playerId: playerId,
             platform: platform,
             testData: nonExistentUrlTestData,
-          );
-          fail('PlatformException not thrown');
-          // ignore: avoid_catches_without_on_clauses
-        } on PlatformException catch (e) {
-          expect(e.message, startsWith('Failed to set source.'));
-        }
-        await tester.pumpLinux();
+          ),
+        );
       },
+      // FIXME(Gustl22): for some reason, the error propagated back from the
+      //  Android MediaPlayer is only triggered, when the timeout has reached,
+      //  although the error is emitted immediately.
+      //  Further, the other future is not fulfilled and then mysteriously
+      //  failing in later tests.
+      skip: isAndroid,
     );
 
     testWidgets('#create and #dispose', (tester) async {
-      await tester.pumpAndSettle();
       await platform.dispose(playerId);
 
       try {
@@ -87,7 +107,6 @@ void main() async {
 
       // Create player again, so it can be disposed in tearDown
       await platform.create(playerId);
-      await tester.pumpLinux();
     });
 
     for (final td in audioTestDataList) {
@@ -99,7 +118,10 @@ void main() async {
             platform: platform,
             testData: td,
           );
-          expect(await platform.getCurrentPosition(playerId), 0);
+          if (!td.isLiveStream) {
+            // Live stream position is not aligned yet.
+            expect(await platform.getCurrentPosition(playerId), 0);
+          }
           final durationMs = await platform.getDuration(playerId);
           expect(
             durationMs != null ? Duration(milliseconds: durationMs) : null,
@@ -111,11 +133,9 @@ void main() async {
               deviation: Duration(milliseconds: td.isVBR ? 100 : 1),
             ),
           );
-          await tester.pumpLinux();
         },
-        // FIXME(gustl22): cannot determine initial duration for VBR on Linux
         // FIXME(gustl22): determines wrong initial position for m3u8 on Linux
-        skip: isLinux && td.isVBR ||
+        skip: !canDetermineDuration(td) ||
             isLinux && td.source == m3u8UrlTestData.source,
       );
     }
@@ -135,7 +155,6 @@ void main() async {
             await platform.stop(playerId);
           }
           // May check native volume here
-          await tester.pumpLinux();
         });
       }
     }
@@ -155,7 +174,6 @@ void main() async {
             await platform.stop(playerId);
           }
           // May check native balance here
-          await tester.pumpLinux();
         });
       }
     }
@@ -175,7 +193,6 @@ void main() async {
             await platform.stop(playerId);
           }
           // May check native playback rate here
-          await tester.pumpLinux();
         });
       }
     }
@@ -189,7 +206,6 @@ void main() async {
       await platform.setPlaybackRate(playerId, 2.0);
       await tester.pumpAndSettle(const Duration(seconds: 2));
       expect(await platform.getCurrentPosition(playerId), 0);
-      await tester.pumpLinux();
     });
 
     for (final td in audioTestDataList) {
@@ -211,7 +227,7 @@ void main() async {
                 onError: seekCompleter.completeError,
               );
           await platform.seek(playerId, const Duration(milliseconds: 22));
-          await seekCompleter.future.timeout(const Duration(seconds: 30));
+          await seekCompleter.future.timeout(_defaultTimeout);
           await onSeekSub.cancel();
           final positionMs = await platform.getCurrentPosition(playerId);
           expect(
@@ -222,7 +238,6 @@ void main() async {
               deviation: const Duration(milliseconds: 1),
             ),
           );
-          await tester.pumpLinux();
         });
       }
     }
@@ -243,7 +258,6 @@ void main() async {
           await platform.stop(playerId);
 
           // May check number of loops here
-          await tester.pumpLinux();
         });
       }
     }
@@ -268,7 +282,6 @@ void main() async {
           // TODO(Gustl22): test if source was released
           expect(await platform.getDuration(playerId), null);
           expect(await platform.getCurrentPosition(playerId), null);
-          await tester.pumpLinux();
         });
       }
     }
@@ -286,7 +299,6 @@ void main() async {
         // Check if position & duration is zero after play & release
         expect(await platform.getDuration(playerId), null);
         expect(await platform.getCurrentPosition(playerId), null);
-        await tester.pumpLinux();
       });
     }
 
@@ -296,9 +308,11 @@ void main() async {
           playerId: playerId,
           platform: platform,
           testData: wavUrl1TestData,
+          // We don't expect the duration event is emitted again,
+          // if the same source is set twice
+          waitForDurationEvent: i == 0,
         );
       }
-      await tester.pumpLinux();
     });
   });
 
@@ -321,42 +335,32 @@ void main() async {
         testWidgets(
           '#durationEvent ${td.source}',
           (tester) async {
-            final eventStream = platform.getEventStream(playerId);
-            final durationCompleter = Completer<Duration?>();
-            final onDurationSub = eventStream
-                .where((event) => event.eventType == AudioEventType.duration)
-                .listen(
-                  (event) => durationCompleter.complete(event.duration),
-                  onError: durationCompleter.completeError,
-                );
+            // Wait for duration before event is emitted.
+            final durationFuture = tester
+                .getDurationFromEvent(
+                  playerId: playerId,
+                  platform: platform,
+                )
+                .timeout(_defaultTimeout);
 
             await tester.prepareSource(
               playerId: playerId,
               platform: platform,
               testData: td,
+              waitForDurationEvent: false,
             );
 
-            if (td.source == wavAsset2TestData.source) {
-              await tester.pumpLinux();
-            }
-
             expect(
-              await durationCompleter.future
-                  .timeout(const Duration(seconds: 30)),
+              await durationFuture,
               (Duration? actual) => durationRangeMatcher(
                 actual,
                 td.duration,
-                deviation: Duration(milliseconds: td.isVBR ? 100 : 1),
+                deviation:
+                    Duration(milliseconds: td.isVBR || isWindows ? 100 : 1),
               ),
             );
-            await onDurationSub.cancel();
-            await tester.pumpLinux();
           },
-          // TODO(gustl22): cannot determine duration for VBR on Linux
-          // FIXME(gustl22): duration event is not emitted for short duration
-          // WAV on Linux (only platform tests, may be a race condition).
-          skip: isLinux && td.isVBR ||
-              isLinux && td.duration! < const Duration(seconds: 5),
+          skip: !canDetermineDuration(td),
         );
       }
     }
@@ -371,19 +375,13 @@ void main() async {
           );
 
           final eventStream = platform.getEventStream(playerId);
-          final completeCompleter = Completer<void>();
-          final onCompleteSub = eventStream
-              .where((event) => event.eventType == AudioEventType.complete)
-              .listen(
-                completeCompleter.complete,
-                onError: completeCompleter.completeError,
-              );
+          final completeFuture = eventStream.firstWhere(
+            (event) => event.eventType == AudioEventType.complete,
+          );
 
           await platform.resume(playerId);
           await tester.pumpAndSettle(const Duration(seconds: 3));
-          await completeCompleter.future.timeout(const Duration(seconds: 30));
-          onCompleteSub.cancel();
-          await tester.pumpLinux();
+          await completeFuture.timeout(_defaultTimeout);
         });
       }
     }
@@ -393,24 +391,18 @@ void main() async {
       for (var i = 0; i < 2; i++) {
         final eventSub = eventStream.listen(null);
         await eventSub.cancel();
-        await tester.pumpLinux();
       }
     });
 
     testWidgets('Emit platform log', (tester) async {
-      final logCompleter = Completer<String>();
-      final logSub = platform
+      final logFuture = platform
           .getEventStream(playerId)
-          .where((event) => event.eventType == AudioEventType.log)
-          .map((event) => event.logMessage)
-          .listen(logCompleter.complete, onError: logCompleter.completeError);
+          .firstWhere((event) => event.eventType == AudioEventType.log)
+          .then((event) => event.logMessage);
 
       await platform.emitLog(playerId, 'SomeLog');
 
-      final log = await logCompleter.future;
-      expect(log, 'SomeLog');
-      await logSub.cancel();
-      await tester.pumpLinux();
+      expect(await logFuture, 'SomeLog');
     });
 
     testWidgets('Emit global platform log', (tester) async {
@@ -431,7 +423,6 @@ void main() async {
       // FIXME: cancelling the global event stream leads to
       // MissingPluginException on Android, if dispose app afterwards
       // await eventStreamSub.cancel();
-      await tester.pumpLinux();
     });
 
     testWidgets('Emit platform error', (tester) async {
@@ -452,7 +443,6 @@ void main() async {
       expect(platformException.code, 'SomeErrorCode');
       expect(platformException.message, 'SomeErrorMessage');
       await eventStreamSub.cancel();
-      await tester.pumpLinux();
     });
 
     testWidgets('Emit global platform error', (tester) async {
@@ -476,7 +466,6 @@ void main() async {
       // FIXME: cancelling the global event stream leads to
       // MissingPluginException on Android, if dispose app afterwards
       // await eventStreamSub.cancel();
-      await tester.pumpLinux();
     });
   });
 }
@@ -486,35 +475,78 @@ extension on WidgetTester {
     required String playerId,
     required AudioplayersPlatformInterface platform,
     required LibSourceTestData testData,
+    bool waitForDurationEvent = true,
+  }) async {
+    final Future<void>? durationFuture;
+
+    if (waitForDurationEvent &&
+        testData.duration != null &&
+        canDetermineDuration(testData)) {
+      // Need to wait for the duration event,
+      // otherwise it gets fired/received after the test has ended,
+      // and therefore then ends up being received in the next test.
+      durationFuture = getDurationFromEvent(
+        playerId: playerId,
+        platform: platform,
+      );
+    } else {
+      durationFuture = null;
+    }
+
+    final eventStream = platform.getEventStream(playerId);
+    final preparedFuture = eventStream
+        .firstWhere(
+          (event) =>
+              event.eventType == AudioEventType.prepared &&
+              (event.isPrepared ?? false),
+        )
+        .timeout(_defaultTimeout);
+
+    Future<void> setSource(Source source) async {
+      if (source is UrlSource) {
+        return platform.setSourceUrl(playerId, source.url);
+      } else if (source is AssetSource) {
+        final cachePath = await AudioCache.instance.loadPath(source.path);
+        return platform.setSourceUrl(playerId, cachePath, isLocal: true);
+      } else if (source is BytesSource) {
+        return platform.setSourceBytes(playerId, source.bytes);
+      } else {
+        throw 'Unknown source type: ${source.runtimeType}';
+      }
+    }
+
+    // Need to await the setting the source to propagate immediate errors.
+    final setSourceFuture = setSource(testData.source);
+
+    // Wait simultaneously to ensure all errors are propagated through the same
+    // future.
+    await Future.wait([setSourceFuture, preparedFuture]);
+    if (durationFuture != null) {
+      await durationFuture;
+    }
+  }
+
+  Future<Duration?> getDurationFromEvent({
+    required String playerId,
+    required AudioplayersPlatformInterface platform,
   }) async {
     final eventStream = platform.getEventStream(playerId);
-    final preparedCompleter = Completer<void>();
-    final onPreparedSub = eventStream
-        .where((event) => event.eventType == AudioEventType.prepared)
-        .map((event) => event.isPrepared!)
-        .listen(
-      (isPrepared) {
-        if (isPrepared) {
-          preparedCompleter.complete();
-        }
-      },
-      onError: (Object e, [StackTrace? st]) {
-        if (!preparedCompleter.isCompleted) {
-          preparedCompleter.completeError(e, st);
-        }
-      },
-    );
-    await pumpLinux();
-    final source = testData.source;
-    if (source is UrlSource) {
-      await platform.setSourceUrl(playerId, source.url);
-    } else if (source is AssetSource) {
-      final cachePath = await AudioCache.instance.loadPath(source.path);
-      await platform.setSourceUrl(playerId, cachePath, isLocal: true);
-    } else if (source is BytesSource) {
-      await platform.setSourceBytes(playerId, source.bytes);
+    final durationFuture = eventStream
+        .firstWhere(
+          (event) => event.eventType == AudioEventType.duration,
+        )
+        .then((event) => event.duration);
+    return durationFuture.timeout(_defaultTimeout);
+  }
+
+  Future<void> expectSettingSourceFailure({
+    required Future<void> future,
+  }) async {
+    try {
+      await future;
+      fail('PlatformException not thrown');
+    } on PlatformException catch (e) {
+      expect(e.message, startsWith('Failed to set source.'));
     }
-    await preparedCompleter.future.timeout(const Duration(seconds: 30));
-    await onPreparedSub.cancel();
   }
 }