+++ lib/src/audioplayer.dart
@@ -1,10 +1,12 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:audioplayers/audioplayers.dart';
 import 'package:audioplayers/src/uri_ext.dart';
 import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-import 'package:meta/meta.dart';
+import 'package:path_provider/path_provider.dart';
 import 'package:uuid/uuid.dart';
 
 const _uuid = Uuid();
@@ -55,10 +57,17 @@ class AudioPlayer {
 
   ReleaseMode get releaseMode => _releaseMode;
 
+  /// Auxiliary variable to re-check the volatile player state during async
+  /// operations.
+  @visibleForTesting
+  PlayerState desiredState = PlayerState.stopped;
+
   PlayerState _playerState = PlayerState.stopped;
 
   PlayerState get state => _playerState;
 
+  /// The current playback state.
+  /// It is only set, when the corresponding action succeeds.
   set state(PlayerState state) {
     if (_playerState == PlayerState.disposed) {
       throw Exception('AudioPlayer has been disposed');
@@ -66,7 +75,7 @@ class AudioPlayer {
     if (!_playerStateController.isClosed) {
       _playerStateController.add(state);
     }
-    _playerState = state;
+    _playerState = desiredState = state;
   }
 
   PositionUpdater? _positionUpdater;
@@ -77,8 +86,6 @@ class AudioPlayer {
   final creatingCompleter = Completer<void>();
 
   late final StreamSubscription _onPlayerCompleteStreamSubscription;
-
-  late final StreamSubscription _onSeekCompleteStreamSubscription;
 
   late final StreamSubscription _onLogStreamSubscription;
 
@@ -157,9 +164,6 @@ class AudioPlayer {
         /* Errors are already handled via log stream */
       },
     );
-    _onSeekCompleteStreamSubscription = onSeekComplete.listen((event) async {
-      await _positionUpdater?.update();
-    });
     _create();
     positionUpdater = FramePositionUpdater(
       getPosition: getCurrentPosition,
@@ -168,7 +172,9 @@ class AudioPlayer {
 
   Future<void> _create() async {
     try {
+      await global.ensureInitialized();
       await _platform.create(playerId);
+      // Assign the event stream, now that the platform registered this player.
       _eventStreamSubscription = _platform.getEventStream(playerId).listen(
             _eventStreamController.add,
             onError: _eventStreamController.addError,
@@ -179,6 +185,10 @@ class AudioPlayer {
     }
   }
 
+  /// Play an audio [source].
+  ///
+  /// To reduce preparation latency, instead consider calling [setSource]
+  /// beforehand and then [resume] separately.
   Future<void> play(
     Source source, {
     double? volume,
@@ -187,6 +197,8 @@ class AudioPlayer {
     Duration? position,
     PlayerMode? mode,
   }) async {
+    desiredState = PlayerState.playing;
+
     if (mode != null) {
       await setPlayerMode(mode);
     }
@@ -205,7 +217,7 @@ class AudioPlayer {
       await seek(position);
     }
 
-    await resume();
+    await _resume();
   }
 
   Future<void> setAudioContext(AudioContext ctx) async {
@@ -224,10 +236,13 @@ class AudioPlayer {
   /// If you call [resume] later, the audio will resume from the point that it
   /// has been paused.
   Future<void> pause() async {
+    desiredState = PlayerState.paused;
     await creatingCompleter.future;
-    await _platform.pause(playerId);
-    state = PlayerState.paused;
-    await _positionUpdater?.stopAndUpdate();
+    if (desiredState == PlayerState.paused) {
+      await _platform.pause(playerId);
+      state = PlayerState.paused;
+      await _positionUpdater?.stopAndUpdate();
+    }
   }
 
   /// Stops the audio that is currently playing.
@@ -235,18 +250,29 @@ class AudioPlayer {
   /// The position is going to be reset and you will no longer be able to resume
   /// from the last point.
   Future<void> stop() async {
+    desiredState = PlayerState.stopped;
     await creatingCompleter.future;
-    await _platform.stop(playerId);
-    state = PlayerState.stopped;
-    await _positionUpdater?.stopAndUpdate();
+    if (desiredState == PlayerState.stopped) {
+      await _platform.stop(playerId);
+      state = PlayerState.stopped;
+      await _positionUpdater?.stopAndUpdate();
+    }
   }
 
   /// Resumes the audio that has been paused or stopped.
   Future<void> resume() async {
+    desiredState = PlayerState.playing;
+    await _resume();
+  }
+
+  /// Resume without setting the desired state.
+  Future<void> _resume() async {
     await creatingCompleter.future;
-    await _platform.resume(playerId);
-    state = PlayerState.playing;
-    _positionUpdater?.start();
+    if (desiredState == PlayerState.playing) {
+      await _platform.resume(playerId);
+      state = PlayerState.playing;
+      _positionUpdater?.start();
+    }
   }
 
   /// Releases the resources associated with this media player.
@@ -256,14 +282,22 @@ class AudioPlayer {
   Future<void> release() async {
     await stop();
     await _platform.release(playerId);
-    state = PlayerState.stopped;
+    // Stop state already set in stop()
     _source = null;
   }
 
   /// Moves the cursor to the desired position.
   Future<void> seek(Duration position) async {
     await creatingCompleter.future;
-    return _platform.seek(playerId, position);
+
+    final futureSeekComplete =
+        onSeekComplete.first.timeout(const Duration(seconds: 30));
+    final futureSeek = _platform.seek(playerId, position);
+    // Wait simultaneously to ensure all errors are propagated through the same
+    // future.
+    await Future.wait([futureSeek, futureSeekComplete]);
+
+    await _positionUpdater?.update();
   }
 
   /// Sets the stereo balance.
@@ -315,26 +349,21 @@ class AudioPlayer {
     await source.setOnPlayer(this);
   }
 
-  Future<void> _completePrepared(Future<void> Function() fun) async {
+  /// This method helps waiting for a source to be set until it's prepared.
+  /// This can happen immediately after [setSource] has finished or it needs to
+  /// wait for the [AudioEvent] [AudioEventType.prepared] to arrive.
+  Future<void> _completePrepared(Future<void> Function() setSource) async {
     await creatingCompleter.future;
-    final preparedCompleter = Completer<void>();
-    late StreamSubscription<bool> onPreparedSubscription;
-    onPreparedSubscription = _onPrepared.listen(
-      (isPrepared) async {
-        if (isPrepared) {
-          preparedCompleter.complete();
-          await onPreparedSubscription.cancel();
-        }
-      },
-      onError: (Object e, [StackTrace? stackTrace]) async {
-        if (!preparedCompleter.isCompleted) {
-          preparedCompleter.completeError(e, stackTrace);
-          await onPreparedSubscription.cancel();
-        }
-      },
-    );
-    await fun();
-    await preparedCompleter.future.timeout(const Duration(seconds: 30));
+
+    final preparedFuture = _onPrepared
+        .firstWhere((isPrepared) => isPrepared)
+        .timeout(const Duration(seconds: 30));
+    // Need to await the setting the source to propagate immediate errors.
+    final setSourceFuture = setSource();
+
+    // Wait simultaneously to ensure all errors are propagated through the same
+    // future.
+    await Future.wait([setSourceFuture, preparedFuture]);
 
     // Share position once after finished loading
     await _positionUpdater?.update();
@@ -344,13 +373,24 @@ class AudioPlayer {
   ///
   /// The resources will start being fetched or buffered as soon as you call
   /// this method.
-  Future<void> setSourceUrl(String url) async {
-    _source = UrlSource(url);
+  Future<void> setSourceUrl(String url, {String? mimeType}) async {
+    if (!kIsWeb &&
+        defaultTargetPlatform != TargetPlatform.android &&
+        url.startsWith('data:')) {
+      // Convert data URI's to bytes (native support for web and android).
+      final uriData = UriData.fromUri(Uri.parse(url));
+      mimeType ??= url.substring(url.indexOf(':') + 1, url.indexOf(';'));
+      await setSourceBytes(uriData.contentAsBytes(), mimeType: mimeType);
+      return;
+    }
+
+    _source = UrlSource(url, mimeType: mimeType);
     // Encode remote url to avoid unexpected failures.
     await _completePrepared(
       () => _platform.setSourceUrl(
         playerId,
         UriCoder.encodeOnce(url),
+        mimeType: mimeType,
         isLocal: false,
       ),
     );
@@ -360,10 +400,15 @@ class AudioPlayer {
   ///
   /// The resources will start being fetched or buffered as soon as you call
   /// this method.
-  Future<void> setSourceDeviceFile(String path) async {
-    _source = DeviceFileSource(path);
+  Future<void> setSourceDeviceFile(String path, {String? mimeType}) async {
+    _source = DeviceFileSource(path, mimeType: mimeType);
     await _completePrepared(
-      () => _platform.setSourceUrl(playerId, path, isLocal: true),
+      () => _platform.setSourceUrl(
+        playerId,
+        path,
+        isLocal: true,
+        mimeType: mimeType,
+      ),
     );
   }
 
@@ -372,19 +417,39 @@ class AudioPlayer {
   ///
   /// The resources will start being fetched or buffered as soon as you call
   /// this method.
-  Future<void> setSourceAsset(String path) async {
-    _source = AssetSource(path);
+  Future<void> setSourceAsset(String path, {String? mimeType}) async {
+    _source = AssetSource(path, mimeType: mimeType);
     final cachePath = await audioCache.loadPath(path);
     await _completePrepared(
-      () => _platform.setSourceUrl(playerId, cachePath, isLocal: true),
+      () => _platform.setSourceUrl(
+        playerId,
+        cachePath,
+        mimeType: mimeType,
+        isLocal: true,
+      ),
     );
   }
 
-  Future<void> setSourceBytes(Uint8List bytes) async {
-    _source = BytesSource(bytes);
-    await _completePrepared(
-      () => _platform.setSourceBytes(playerId, bytes),
-    );
+  Future<void> setSourceBytes(Uint8List bytes, {String? mimeType}) async {
+    if (!kIsWeb &&
+        (defaultTargetPlatform == TargetPlatform.iOS ||
+            defaultTargetPlatform == TargetPlatform.macOS ||
+            defaultTargetPlatform == TargetPlatform.linux)) {
+      // Convert to file as workaround
+      final tempDir = (await getTemporaryDirectory()).path;
+      final bytesHash = Object.hashAll(bytes)
+          .toUnsigned(20)
+          .toRadixString(16)
+          .padLeft(5, '0');
+      final file = File('$tempDir/$bytesHash');
+      await file.writeAsBytes(bytes);
+      await setSourceDeviceFile(file.path, mimeType: mimeType);
+    } else {
+      _source = BytesSource(bytes, mimeType: mimeType);
+      await _completePrepared(
+        () => _platform.setSourceBytes(playerId, bytes, mimeType: mimeType),
+      );
+    }
   }
 
   /// Set the PositionUpdater to control how often the position stream will be
@@ -428,13 +493,12 @@ class AudioPlayer {
     // First stop and release all native resources.
     await release();
 
-    state = PlayerState.disposed;
+    state = desiredState = PlayerState.disposed;
 
     final futures = <Future>[
       if (_positionUpdater != null) _positionUpdater!.dispose(),
       if (!_playerStateController.isClosed) _playerStateController.close(),
       _onPlayerCompleteStreamSubscription.cancel(),
-      _onSeekCompleteStreamSubscription.cancel(),
       _onLogStreamSubscription.cancel(),
       _eventStreamSubscription.cancel(),
       _eventStreamController.close(),