+++ test/global_audioplayers_test.dart
@@ -33,13 +33,15 @@ void main() {
     /// If using [AVAudioSessionCategory.playAndRecord] the audio will come from
     /// the earpiece unless [AVAudioSessionOptions.defaultToSpeaker] is used.
     test('set AudioContext', () async {
-      await globalScope.setAudioContext(const AudioContext());
-      final call = globalPlatform.popLastCall();
+      await globalScope.setAudioContext(AudioContext());
+      var call = globalPlatform.popCall();
+      expect(call.method, 'init');
+      call = globalPlatform.popLastCall();
       expect(call.method, 'setGlobalAudioContext');
       expect(
         call.value,
-        const AudioContext(
-          android: AudioContextAndroid(
+        AudioContext(
+          android: const AudioContextAndroid(
             isSpeakerphoneOn: false,
             audioMode: AndroidAudioMode.normal,
             stayAwake: false,